superskills/vision/requirements.txt
# Core dependencies (from main requirements.txt)
google-genai>=0.2.0
Pillow>=10.0.0

# Vision-specific
pyautogui>=0.9.54
numpy>=1.24.0  # For image comparison
